feat:: calendar service에 custom exception 적용
- 불필요한 예외처리 및 repositoryt 단의 코드 삭제
- 불필요한 dto 필드 리팩토링
