+++ src/home/calendar/calendar.repository.ts
@@ -1,4 +1,4 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { CalendarEntity } from 'src/entities/calendar.entity';
 import {
   DataSource,
@@ -6,7 +6,6 @@ import {
   MoreThanOrEqual,
   Repository,
 } from 'typeorm';
-import { UpdateCalendarDataRequestDto } from './dto/update-calendar-data-request.dto';
 
 @Injectable()
 export class CalendarRepository extends Repository<CalendarEntity> {
@@ -48,32 +47,5 @@ export class CalendarRepository extends Repository<CalendarEntity> {
     });
 
     return await this.save(calendarData);
-  }
-
-  async updateCalendarData(
-    calendarId: number,
-    requestDto: UpdateCalendarDataRequestDto,
-  ): Promise<CalendarEntity> {
-    const updateData = {
-      ...requestDto,
-      startDate: requestDto.startDate
-        ? new Date(requestDto.startDate)
-        : undefined,
-      endDate: requestDto.endDate ? new Date(requestDto.endDate) : undefined,
-    };
-
-    const updated = await this.update({ id: calendarId }, updateData);
-
-    if (updated.affected === 0) {
-      throw new InternalServerErrorException('업데이트에 실패했습니다.');
-    }
-
-    return await this.findOne({ where: { id: calendarId } });
-  }
-
-  async deleteCalendarData(calendarId: number): Promise<boolean> {
-    const deleted = await this.softDelete({ id: calendarId });
-
-    return deleted.affected ? true : false;
   }
 }
