+++ src/home/calendar/calendar.service.ts
@@ -1,8 +1,4 @@
-import {
-  Injectable,
-  InternalServerErrorException,
-  NotFoundException,
-} from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CalendarRepository } from './calendar.repository';
 import { GetDailyCalendarDataResponseDto } from './dto/get-calendar-data-response-dto';
@@ -14,6 +10,7 @@ import { DeleteCalendarDataResponseDto } from './dto/delete-calendar-data-respon
 import { GetAcademicScheduleDataResponseDto } from './dto/get-academic-schedule-response.dto';
 import { GetBannerImageUrlResponseDto } from './dto/get-banner-images-response.dto';
 import { FileService } from 'src/common/file.service';
+import { throwKukeyException } from 'src/utils/exception.util';
 
 @Injectable()
 export class CalendarService {
@@ -101,10 +98,6 @@ export class CalendarService {
       isAcademic,
     );
 
-    if (!calendarData) {
-      throw new InternalServerErrorException('행사/일정 생성에 실패했습니다.');
-    }
-
     return new CreateCalendarDataResponseDto(calendarData);
   }
 
@@ -117,13 +110,27 @@ export class CalendarService {
     });
 
     if (!calendarData) {
-      throw new NotFoundException('행사/일정 정보가 없습니다.');
+      throwKukeyException('CALENDAR_NOT_FOUND');
     }
 
-    return await this.calendarRepository.updateCalendarData(
-      calendarId,
-      requestDto,
+    const updateData = {
+      ...requestDto,
+      startDate: requestDto.startDate
+        ? new Date(requestDto.startDate)
+        : undefined,
+      endDate: requestDto.endDate ? new Date(requestDto.endDate) : undefined,
+    };
+
+    const updated = await this.calendarRepository.update(
+      { id: calendarId },
+      updateData,
     );
+
+    if (updated.affected === 0) {
+      throwKukeyException('CALENDAR_UPDATE_FAILED');
+    }
+
+    return new UpdateCalendarDataResponseDto(true);
   }
 
   async deleteCalendarData(
@@ -134,14 +141,15 @@ export class CalendarService {
     });
 
     if (!calendarData) {
-      throw new NotFoundException('행사/일정 정보가 없습니다.');
+      throwKukeyException('CALENDAR_NOT_FOUND');
     }
 
-    const isDeleted =
-      await this.calendarRepository.deleteCalendarData(calendarId);
+    const deleted = await this.calendarRepository.softDelete({
+      id: calendarId,
+    });
 
-    if (!isDeleted) {
-      throw new InternalServerErrorException('삭제에 실패했습니다.');
+    if (deleted.affected === 0) {
+      throwKukeyException('CALENDAR_DELETE_FAILED');
     }
 
     return new DeleteCalendarDataResponseDto(true);
